Make HIGH fan speed detection more robust
I don't know why/how but my AC just had a period where it wouldn't send
a HIGH speed the normal way.

Here I up it via the remote:

[20:07:22.395][I][mhi_ac_ctrl:372]: Fan speed 2, changed: 0, raw: 12 / 88
[20:07:22.395][I][mhi_ac_ctrl:372]: Fan speed 6, changed: 1, raw: 12 / c8
[20:07:22.395][D][climate:427]: 'AC Harm' - Sending state:
[20:07:22.395][D][climate:430]:   Mode: FAN_ONLY
[20:07:22.395][D][climate:432]:   Action: FAN
[20:07:22.395][D][climate:435]:   Fan Mode: HIGH
[20:07:22.395][D][climate:450]:   Current Temperature: 19.75°C
[20:07:22.395][D][climate:457]:   Target Temperature: 18.00°C
[20:07:22.395][I][mhi_ac_ctrl:372]: Fan speed 6, changed: 0, raw: 12 / c8
[20:07:22.410][I][mhi_ac_ctrl:372]: Fan speed 6, changed: 0, raw: 12 / c8
[20:07:22.416][I][mhi_ac_ctrl:372]: Fan speed 6, changed: 0, raw: 12 / c8
[20:07:22.431][I][mhi_ac_ctrl:372]: Fan speed 6, changed: 0, raw: 12 / c8

The first raw value is DB1, the second DB6.

After a while I was testing some more (no ESP reboots in between):

[20:12:10.444][I][mhi_ac_ctrl:372]: Fan speed 2, changed: 0, raw: 12 / 88
[20:12:10.444][I][mhi_ac_ctrl:372]: Fan speed 6, changed: 1, raw: 16 / 88
[20:12:10.444][D][climate:427]: 'AC Harm' - Sending state:
[20:12:10.444][D][climate:430]:   Mode: FAN_ONLY
[20:12:10.445][D][climate:432]:   Action: FAN
[20:12:10.445][D][climate:435]:   Fan Mode: HIGH
[20:12:10.445][D][climate:450]:   Current Temperature: 18.00°C
[20:12:10.445][D][climate:457]:   Target Temperature: 18.00°C
[20:12:10.445][I][mhi_ac_ctrl:372]: Fan speed 6, changed: 0, raw: 16 / 88

Weird, huh? I was testing the ESPHome 2025.11 upgrade and noticed that
my fan speed in HA did not go higher than medium when I was using the
remote. Which is in line with what [1] says (the code didn't read DB6 at
that point). All of a sudden (to me) behaviour changes. This makes the
code more robust. In the case that the behaviour of [1] is going on, we
won't have auto fan speed detection but at least the others will
display.

[1] https://github.com/absalom-muc/MHI-AC-Trace/blob/71d33871540fb190d76835627bdbc3bfcb0848b2/SPI.md